LayoutTests/platform/ios-sim-deprecated/fast/forms/file/file-style-inheritance-expected.txt

Tests internal components of a file upload control correctly inherit the style of the container input element.



FAIL document.defaultView.getComputedStyle(button1, null).getPropertyValue("font-size") should be 30px. Was 11px.
PASS document.defaultView.getComputedStyle(document.getElementById("darkFile"), null).getPropertyValue("background-color") is "rgba(0, 0, 0, 0)"
PASS document.defaultView.getComputedStyle(document.getElementById("darkFile"), null).getPropertyValue("color") is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE
